+++ extensions/brainpylib/custom_op/__init__.py
@@ -1,0 +1,3 @@
+from cpu import *
+from gpu import *
+from regis_op import register_op
