--- extensions/brainpylib/operator/__init__.py
@@ -1,2 +1,0 @@
-from cpu_op import *
-from gpu_op import *
